Expand the usages of prebuilt_any in prebuilt_etc module gen
Currently, prebuilt_any is only used as a fallback behavior when the
generated prebuilt_root module tries to set the `Dsts` property, given
that prebuilt_root module type is no longer allowed to set `Dsts`
property with https://r.android.com/3540101.

This change expands the usages of the fallback behavior to when the
generated prebuilt_root module tries to set the `Relative_install_path`
property, given that this property is not also available to
prebuilt_root module type.

Test: m nothing --no-skip-soong-tests
Change-Id: I9344a8cb3e26bedbd0217a5344ac92b8def68554
Signed-off-by: HeroBuxx <me@herobuxx.me>